Added button to main page and changed sign in layout
Added a button to the main page that directs user to the sign on page. Currently sits above the mission paragraphs will change later as I mess with the layout of that page.

Changed the layout and style of the sign in page. Added an image of LHU logo to display above the form and a little paragraph under the form.